1. Adds functionality to disable buttons based on the currently logged in user's employee type.
2. Adds functionality to display the currently logged in user's information.
3. Fixes several ui and code bugs.

+++ code/java_project/src/edu/westga/cs3212/inventory_manager/viewmodel/InventoryViewModel.java
@@ -3,7 +3,9 @@ package edu.westga.cs3212.inventory_manager.viewmodel;
 import java.util.ArrayList;
 import java.util.List;
 
+import edu.westga.cs3212.inventory_manager.Main;
 import edu.westga.cs3212.inventory_manager.model.Component;
+import edu.westga.cs3212.inventory_manager.model.Constants;
 import edu.westga.cs3212.inventory_manager.model.Item;
 import edu.westga.cs3212.inventory_manager.model.Product;
 import edu.westga.cs3212.inventory_manager.model.local_impl.LocalComponentInventory;
@@ -165,6 +167,11 @@ public class InventoryViewModel {
 	 *                inventory.
 	 */
 	public void removeProduct() {
+		if (ProductInventory.checkAnyProductUsesComponent(
+				this.selectedComponent.get().getID())) {
+			throw new IllegalArgumentException(
+					Constants.ERROR_COMPONENT_IN_USE_BY_PRODUCT);
+		}
 		ProductInventory.deleteProduct(this.selectedProduct.get().getID());
 	}
 
@@ -183,6 +190,15 @@ public class InventoryViewModel {
 	 */
 	public void produceProduct(Product selectedProduct, int quantity) {
 		ProductInventory.produceProduct(selectedProduct.getID(), quantity);
+	}
+	
+	/**
+	 * Determines if the user is a manager.
+	 * 
+	 * @return true if the user is a manager, false otherwise
+	 */
+	public boolean isManager() {
+		return Main.isLoggedInEmployeeManager();
 	}
 	
 	public ObservableList<Component> searchComponents(String searchString) {
